# Optimize the copy that occurs in ets:lookup_element