@sverker 2022-09-27
Thanks for the contribution.

I pushed a memory leak fix in `copy_struct_in_order_x` to your `copy` branch.

Question: Did you contemplate looping over `copy_struct` for each element instead of implementing `copy_struct_in_order_x`?