@RobinMorisset 2022-09-29
Thank you for catching the memory leak, I'm sorry for missing it.

I did not think to just loop over copy_struct_x, I agree that it sounds much simpler, I'll try that right now.