@sverker 2022-09-29
> Thank you for catching the memory leak, I'm sorry for missing it.

I just ran `ets_SUITE` with `debug` compiled emulator.
```
> cd $ERL_TOP
> (cd erts/emulator && make debug)
> bin/cerl -debug
```
For more development tips and tools see https://www.erlang.org/doc/tutorial/debugging.html