@RobinMorisset 2022-10-03
It turns out that just calling copy_struct for each field of the tuple is actually very tricky.
The problem I found is lists: the beginning of the list is not guaranteed by copy_struct to be at the beginning of the space allocated for the field. For example, the tuple `{x, "AA"}` has the following representation if I store it in ETS with repeated calls to copy_struct:
```
        OTHER at 0x1465ad398, 0x80
	IMMED1 at 0x1465ad3a0, 628811 (atom x)
	LIST at 0x1465ad3a8, 0x1465ad3c0 (first cons)
	IMMED1 at 0x1465ad3b0, 1055 (A)
	IMMED1 at 0x1465ad3b8, 59 (NIL)
	IMMED1 at 0x1465ad3c0, 1055 (A)
	LIST at 0x1465ad3c8, 0x1465ad3b0 (second cons)
```
Note that obj->tpl[2] does point to "AA", but it starts near the end of its allocated space (0x1465ad3c0) and not the beginning (0x1465ad3b0). This results in db_field_size getting the wrong size for the field, and only part of it is copied.
I don't see how to fix this, this whole patch relies on:
- each field is allocated as a block
- the tuple at the top has pointers to the beginning of each block

I'll document it in an updated commit message/comments.

As for performance, I've figured how to do benchmarking on our internal workload, I'll use it to verify that my original approach does not add a huge cost to ets:insert.