@sverker 2022-10-03
Oh, I didn't think about the putting-things-last optimization in copy_struct. Hm....

_I'm sorry if I persist skeptical of your quite impressive `copy_struct_in_order_x()`, but as a maintainer I'm a bit stickler about keeping the code volume down if possible._

How about something like this, where we only give special treatment to the list case and delegate the rest to standard copy_struct:

```
Eterm copy_ets_element(Eterm obj, int sz, Eterm **hpp, ErlOffHeap *off_heap)
{
    if (is_list(obj) && is_immed(CAR(list_val(obj)))) {
        /* copy_struct() would put this last, but we need the top term to be first in block */
        Eterm* src = list_val(obj);
        Eterm* dst = *hpp;

        CAR(dst) = CAR(src);
        *hpp += 2;
        CDR(dst) = copy_struct(CDR(src), sz-2, hpp, off_heap);
        return make_list(dst);
    }
    else {
        return copy_struct(obj, sz, hpp, off_heap);
    }
}
```
There are other cases in `copy_struct()` where things are put last in the block (I searched for `hbot`), but they are all leaf terms which means they will also be the only term and therefor also be first in the block (I think).