@RobinMorisset 2022-10-04
Thanks, I tried and it seems to work! I'll do more testing and benchmarking and comment again.