@sverker 2023-02-09
Fixed bug causing failed debug assert in `ets_SUITE:lookup_element_mult`.