@sverker 2023-02-21
@RobinMorisset 
I've fixed some problems with `ets:update_element` and the empty tuple (which is always a literal).
Now I'm satisfied and will let it run at least one night before squash and merge.