@RobinMorisset 2023-02-21
@sverker Thank you very much for these patches, and sorry for dropping the ball on this. I've spent the last couple of days trying to track down a bug in this PR that only triggers in one of our internal tests, and unfortunately it is still present. So please don't merge the PR right now (but I'd love it if you find any simple testcase for the bug through your own internal testing).