@RobinMorisset 2023-02-21
@sverker Thank you very much for this, and sorry for dropping the ball on this PR. I had been investigating a bug found in it by our internal tests for the past few days, and it turned out to be caused by the empty tuple bug that you just fixed.